requirements.txt
python-dotenv
fastapi>=0.95.1,<0.96.0
pydantic>=1.10.7,<1.11.0
uvicorn>=0.21.1,<0.22.0
azure-core>=1.26.4,<1.27.0
azure-identity>=1.12.0,<1.13.0
azure-mgmt-storage>=21.0.0,<21.1.0